Added SideBarUser with basic styles. Redefined all its routes
