+++ client/src/App.jsx
@@ -6,7 +6,6 @@ import Nav from "./clientComponents/Nav/Nav.jsx";
 import Detail from "./clientViews/Detail/Detail.jsx";
 import Home from "./clientViews/Home/Home.jsx";
 import NotFound from "./clientViews/NotFound/NotFound.jsx";
-import MyProfile from "./clientViews/MyProfile/MyProfile.jsx";
 import Footer from "./clientComponents/Footer/Footer.jsx";
 import EditForm from "./adminViews/EditFood/EditFood";
 import AdminPanel from "./adminViews/AdminPanel/AdminPanel";
@@ -17,7 +16,7 @@ import PaymentStatus from "./clientViews/PaymentStatus/PaymentStatus";
 import OrderDetail from "./adminViews/OrderDetail/OrderDetail";
 import UserOrder from "./clientViews/UserOrder/UserOrder";
 import UserOrderDetail from "./clientViews/UserOrderDetail/UserOrderDetail";
-import MyProfileUser from "./clientViews/MyProfileUser/MyProfileUser";
+import MyProfile from "./clientViews/MyProfile/MyProfile.jsx";
 
 function App() {
   const location = useLocation();
@@ -47,9 +46,10 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/viandas" element={<Viandas />} />
-        <Route path="/myprofile/*" element={<MyProfile />} />
-        <Route path="/userorder" element={ <UserOrder /> } />
-        <Route path="/myprofileuser" element={<MyProfileUser />} />
+        <Route path="/micuenta/misdatos" element={<MyProfile />} />
+        <Route path="/micuenta/misordenes" element={<UserOrder />} />
+        <Route path="/micuenta/misordenes/detalle/:id" element={<UserOrderDetail />} />
+
         <Route path="/detail/:id" element={<Detail />} />
         <Route
           path="/admin/*"
