--- client/src/clientViews/MyProfile/MyProfile.module copy.css
@@ -1,19 +1,0 @@
-.mainContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.rowContainer {
-    margin: 15px;
-}
-
-.input {
-    width: 300px;
-}
