--- client/src/clientViews/MyProfileUser/MyProfileUser.module.css
@@ -1,19 +1,0 @@
-.mainContainer {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.rowContainer {
-    margin: 15px;
-}
-
-.input {
-    width: 300px;
-}
