+++ client/src/clientViews/UserOrderDetail/UserOrderDetail.module.css
@@ -1,7 +1,13 @@
-.orderdiv{
+.mainContainer {
+  display: grid;
+  grid-template-columns: 1fr 5fr; 
+  background-color: white;
+}
+
+.orderdiv {
   padding: 5px 0px 5px 10px;
 }
-.destable{
+.destable {
   border-collapse: collapse;
   min-width: 1111px;
   box-shadow: 0 0 20px rgba(255, 255, 255, 0.15);
@@ -9,7 +15,7 @@
 
 .destable thead tr {
   background-color: var(--accentColor2);
-  color: #ffffff; 
+  color: #ffffff;
   text-align: middle;
   height: 40px;
   font-size: 18px;
@@ -20,14 +26,14 @@
 }
 
 .destable tbody tr {
-  border-bottom: 1px solid #dddddd; 
+  border-bottom: 1px solid #dddddd;
   background: white;
-  } 
-  .destable tbody tr:nth-of-type(even) { 
-      background-color: #f3f3f3; 
-  }
-  .destable tbody tr:last-of-type {
-      border-bottom: 2px solid #009879;
+}
+.destable tbody tr:nth-of-type(even) {
+  background-color: #f3f3f3;
+}
+.destable tbody tr:last-of-type {
+  border-bottom: 2px solid #009879;
 }
 .star {
   color: gray;
